Move useRefreshToken to api-hooks
